first test for all times views done + might change structure of functions

+++ kafka-stream/src/main/scala/org/esgi/project/streaming/models/Likes.scala
@@ -2,7 +2,7 @@ package org.esgi.project.streaming.models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Likes (_id: Int, score: Double)
+case class Likes (_id: Long, score: Double)
 
 object Likes {
   implicit val format: OFormat[Likes] = Json.format[Likes]
