+++ kafka-stream/src/main/scala/org/esgi/project/streaming/models/Views.scala
@@ -2,7 +2,7 @@ package org.esgi.project.streaming.models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Views(_id: Int, title: String, viewsCategory: String)
+case class Views(_id: Long, title: String, viewsCategory: String)
 
 object Views {
   implicit val format: OFormat[Views] = Json.format[Views]
